getting data to show up on journal with json data
