--- Scripts/api/JournalData.json
@@ -1,71 +1,0 @@
-{
-
-"users": [
-            {
-        "name": "Autumn Fotopoulos",
-        "email": "acarte94@gmail.com"
-            }
-          ],
-
-     "entry": [1,
-        {
-        "id": 1,
-        "imageURL": "https://media.giphy.com/media/jPBlSAfeQts9goDVC4/giphy.gif",
-        "date": "02/17/2021",
-        "concept": "Terminal use",
-        "description": "Learned how to use the Terminal, learned alot of commands.",
-        "mood": "Hype"
-    },
-
-{       "id": 2,
-         "imageURL": "https://media.giphy.com/media/jPBlSAfeQts9goDVC4/giphy.gif",
-        "date": "02/18/2021",
-        "concept": "Github",
-        "entry": "Learned how to use Github and Git, first practice with flexbox",
-        "mood": "Chill"
-},
-
-    {
-         "id": 3,
-         "imageURL": "https://media.giphy.com/media/Q8IdwVBZgv46X3vdhZ/giphy.gif",
-        "date": "02/19/2021",
-        "concept": "Github",
-        "entry": "Learning more about Github and pushing up/ pulling down repos.",
-        "mood": "Hype"
-    },
-    {
-        "id": 4,
-        "imageURL": "https://media.giphy.com/media/6tY4W4s9u7mipdv6ba/giphy.gif",
-        "date": "02/20/2021",
-        "concept": "Group Project- HTML/CSS",
-        "entry": "Magical Mecurians group project making a travel page with HTML and CSS.",
-        "mood": "Hype"
-    },
-    {
-        "id": 5,
-        "imageURL": "https://media.giphy.com/media/6tY4W4s9u7mipdv6ba/giphy.gif",
-        "date": "02/21/2021",
-        "concept": "Javascript",
-        "entry": "Intro into Javascript.",
-        "mood": "Chill"
-    },
-    {
-        "id": 6,
-        "imageURL": "https://media.giphy.com/media/6tY4W4s9u7mipdv6ba/giphy.gif",
-        "date": "03/1/2021",
-        "concept": "JS/HTML/CSS",
-        "entry": "Working on group project, Awesome Apricots, making our first project to be automated javascript to go to the DOM.",
-        "mood": "Hype"
-    },
-    {
-        "id": 7,
-        "imageURL" :"https://media.giphy.com/media/A6aHBCFqlE0Rq/giphy.gif",
-        "date": "03/3/2021",
-        "concept": "TJS/HTML/CSS",
-        "entry": "Project presentations, self assesment- missed 2.",
-        "mood": "Chill"
-    }
-
-]
-
-}
