--- Scripts/main.js
@@ -1,3 +1,0 @@
-import { journalList } from "./JournalEntryList.js";
-
-journalList();
