+++ Styles/main.css
@@ -52,7 +52,7 @@ main {
     
 }
 
-.journalList{
+.entryList{
     display: flex;
     flex-wrap: wrap;
     justify-content: space-evenly;
