+++ src/index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Autumn's Journal</title>
-    <link rel="stylesheet" href="styles/main.css">
+    <link rel="stylesheet" href="./styles/main.css">
 </head>
 <body>
     
@@ -17,7 +17,7 @@
 
 <main>
 
-    <section id= "journalList"></section>
+    <section id= "entryList"></section>
     <section id= "journalEntry"></section>
 
     <section class = "journalForm">
